requirements.txt
astroid==2.0
beautifulsoup4==4.6.0
click==6.7
elasticsearch==6.3.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
micawber==0.3.5
pkg-resources==0.0.0
Pygments==2.2.0
pylint==2.0.0
python-slugify==1.2.5
six==1.11.0
SQLAlchemy==1.2.7
timeago==1.0.8
typed-ast==1.1.0
Unidecode==1.0.22
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2
